Add required context to rerouting action
Route action now gets the current application state (= router + location)
as the payload parameter. Thanks to this improvement, the route action
can decide how to update/change the state next. For example, routing params
set by user are not suddendly forgotten, and editing a location in the
summary page does not push a new entry to history. UI stays in sync with
presented routes and browsing history beecomes predictable.