Extend payload for endpoind actions to include react-router context
Endpoint change is a very important action in app architecture -
a change usually launches a route search, which may or may not
result to a transition to a new page. The input data required
for such transition logic is not available in any fluxible store,
but in the react-router context. The changes in this commit
extend the payload to include react router and location, so that
proper actions can be carried out.