Update history when nececessary while opening nearby/favourites tabs in mobile
This fixes the bug that clicking browser's back arrow from the mentioned
tabs usually took the user away from the whole application.
Note that switching between full screen tabs replaces the history; it
does not accumulate a chain of tab switches into history, which apparently
would be useless for navigation purposes and mainly and annoying.