Change attribution when changing map mode
Fixes #108

I looked at the Github repo of react-leaflet and found the following
issue: https://github.com/PaulLeCam/react-leaflet/issues/102

There people discussed a related problem and I took some inspiration
from that.

The solution was to add the `attribution` prop to TileLayer component
and not use the prefix in AttributionControl. In retrospect, this
seem like the better place anyway.