fix: handle uncaught relay parsing error globally
When relay 1.6.0 encounters a GraphQL parsing error it doesn't seem to return
a failed promise or at least a synchronous exception but it schedules
an exception to be thrown later:

https://github.com/facebook/relay/blob/v1.6.0/packages/react-relay/classic/util/throwFailedPromise.js

For example, you can provoke a parsing error by setting the triangle factors of the
bike routing in the URL to add up to more than 1.0.

When the page is rendered with `react-isomorphic-relay-router` then this
leads to a server crash as the exception is thrown completely outside
the ordinary request/response cycle and there is no handler for it.