fix: read location state from location state
In the past, existence of key location.type === 'CurrentLocation' was used
(for the sake of maximum fuzziness) to decide if locationing was already attempted.
So, recent addition of that key to location state at store initialization
broke geolocationing process. Now, the decision is made using the location status key,
which somehow makes sense.