workaround: define bbnavi-* proxy themes
Currently, the theme loading logic depends on a 1-to-1 relationship between
configs (app/configuration/config.*.js) and themes (sass/themes/*).

Because we recently introduced the config.bbnavi-*.js configs, which are all
supposed to use the `bbnavi` theme, we define "proxy themes" here, which just
import the `bbnavi` theme.

The proper fix will be cleaning up the theme loading logic:

- use config.CONFIG instead of process.env.CONFIG
	- https://gitlab.tpwd.de/tpwd/bb-navi/digitransit-ui/-/blob/9aec60a4a33992ed9ab6b383778a7a279b8c5e1c/build/contextHelper.js#L52
- rename config.CONFIG to config.THEME